Moves NormalizedURIRepo and related repos to Shoebox Module
