--- kifi-backend/modules/shoebox/test/com/keepit/common/store/FakeShoeboxStoreModule.scala
@@ -1,11 +1,0 @@
-package com.keepit.common.store
-
-import com.google.inject.{Singleton, Provides}
-import com.keepit.common.analytics.reports.{InMemoryReportStoreImpl, ReportStore}
-
-case class FakeShoeboxStoreModule() extends FakeStoreModule {
-
-  @Provides @Singleton
-  def reportStore(): ReportStore = new InMemoryReportStoreImpl()
-
-}
